removed broken teacher dashboard UI tests for now
