+++ Student/StudentUITests/TestsFromCore.swift
@@ -29,7 +29,6 @@ class StudentActAsUserTests: ActAsUserTests {}
 class StudentAnnouncementTests: AnnouncementTests {}
 class StudentTodoListTests: TodoListTests {}
 class StudentDashboardTests: DashboardTests {}
-class StudentDashboardEmptyTests: DashboardEmptyTests {}
 class StudentProfileTests: ProfileTests {}
 class StudentPageTests: PageTests {}
 class StudentLoginTests: LoginTests {}
